Se agrega el crud para los insumos

+++ src/main/java/com/backend/insumo_servicio/repository/TipoInsumoRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.backend.insumo_servicio.model.TipoInsumo;
 
-public interface TipoInsumoRepository extends JpaRepository<TipoInsumo, Integer> {
+public interface TipoInsumoRepository extends JpaRepository<TipoInsumo, Long> {
 }
